Add base entity and data classes (#3)
* added the ImageLoader

* changed the ImageLoader to AssetsLoader

* submit change

* here

* wip of entity,  rewritten the Eslint to allow for of

* entity system can be initialized

* Test this is a test change 

This is a test change to see if the remote works

* removed entity sprite manual resize

I added the sprite resize and also the dependencies to the assets loader. it now pass a texture file.

Although I did implemented in the data structure an filename and a index parameters. It's similar to how RM handle filename when getting characters.

* Added debug to the gitignore and tested the new change

* changed line ending

* Fixed the linting errors

* Fix lline endings

* Remove dupe tsconfig property

* Fix remaining eslint errors

Co-authored-by: inc0der <inc0der.dev@gmail.com>